% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT2.R
\name{STAT2}
\alias{STAT2}
\title{Launch 'STAT2' Interface}
\usage{
STAT2()
}
\value{
Nothing
}
\description{
STAT2() loads interactive user interface built using R 'shiny'.
}
\details{
Start STAT2

The interactive user interface is to provide an easy way for basic statistical analysis and downloading plots.
}
\examples{
if(interactive()){
library(rmarkdown)
STAT2()
}
}
\keyword{STAT2}
